Updated verision as of 5/6
Can now store orders in database. Need to insert schedule permission for admin , which Jon S is working on. John H maybe you can work on how return the order from the database,  if a user wants to view a previous order.  Ill work on the delete option tomorrow.  In addtion, Ill look into whether we should keep track of table # or not and see if we have the time. Lastly, I got rid of all the foriegn keys in the database tables because I need to make changes to some of the tables. We can fix the foriegn keys once we get all the code working and have all the table columns the way we want.